hotfix: show different dashboard for inspector user
